Extract source product status mapping
