+++ src/api/services/construction_source_product_status.py
@@ -1,0 +1,12 @@
+from src.core.construction.vocabulary import ConstructionMethodStatus
+
+
+def source_status_to_method_status(status: str) -> ConstructionMethodStatus:
+    if status == "READY":
+        return ConstructionMethodStatus.READY
+    if status == "DEGRADED":
+        return ConstructionMethodStatus.DEGRADED
+    return ConstructionMethodStatus.BLOCKED
+
+
+__all__ = ["source_status_to_method_status"]
