+++ src/api/services/construction_treasury_source_context.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 
 from src.core.common.canonical import hash_canonical_payload
-from src.api.services.construction_liquidity_source_context import source_status_to_method_status
+from src.api.services.construction_source_product_status import source_status_to_method_status
 from src.core.construction.models import AuthoritativeCurrencyOverlayContext
 from src.core.dpm_source_context import (
     DpmCoreExternalCurrencyExposureResponse,
